+++ docs/changes.rst
@@ -9,6 +9,7 @@ Released on ? 2019.
 * Fixed Celery systemd unit example.
 * Fixed notifications from http repositories with login.
 * Fixed race condition in editing source string for monolingual translations.
+* Include output of failed addon execution in the logs.
 * Localization updates.
 
 weblate 3.5